# Implement some best practices, fix error in static_ips test & avoid excessive regex calls when line is not spruce line