--- a/op_load.go
+++ b/op_load.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 
 	"github.com/geofffranks/simpleyaml"
-	"github.com/geofffranks/spruce/log"
+	log "github.com/geofffranks/spruce/log"
 	"github.com/starkandwayne/goutils/ansi"
 	"github.com/starkandwayne/goutils/tree"
 )
@@ -63,7 +63,7 @@ func (LoadOperator) Run(ev *Evaluator, args []*Expr) (*Response, error) {
 		s, err := v.Reference.Resolve(ev.Tree)
 		if err != nil {
 			log.DEBUG("     [%d]: resolution failed\n    error: %s", i, err)
-			return nil, fmt.Errorf("Unable to resolve `%s`: %s", v.Reference, err)
+			return nil, fmt.Errorf("unable to resolve `%s`: %s", v.Reference, err)
 		}
 
 		switch s.(type) {