--- a/op_stringify.go
+++ b/op_stringify.go
@@ -1,7 +1,7 @@
 package spruce
 
 import (
-	"github.com/geofffranks/spruce/log"
+	log "github.com/geofffranks/spruce/log"
 	"github.com/geofffranks/yaml"
 	fmt "github.com/starkandwayne/goutils/ansi"
 	"github.com/starkandwayne/goutils/tree"