… merge.go
… op_calc.go
… op_cartesian_product.go
… op_concat.go
… op_file.go
… op_grab.go
… op_inject.go
… op_ips.go
… op_join.go
… op_keys.go
… op_negate.go
… op_prune.go
… op_shuffle.go
… op_sort.go
… op_static_ips.go
… op_vault.go
… operator.go
… operator_test.go
… vault_test.go